Non-stationary Non-overlapping Allan Variance

Calculation of the theoritical Non-overlapping Allan Variance for constant-mean non-stationary time series data

NOAV(n, covmat)

Arguments

n
A integer indicating the length of each vector of consecutive observations considered for the average.

covmat
A matrix indicating the T-by-T covariance matrix of the time series with length T.

Value

A field<numeric> that is the theoritical Non-overlapping Allan Variance for constant-mean non-stationary time series data

Details

Calculate Non-overlapping Allan Variance based on the defination on "Xu, Haotian, et al. "A Study of the Allan Variance for Constant-Mean Non-Stationary Processes." arXiv preprint arXiv:1702.07795 (2017)". n is an integer larger than 1 and smaller than \(floor\left(log_2 \left(dim\left(covmat\right)[1]\right)\right)-1\)

Examples

library(simts)
set.seed(999)
x = gen_gts(AR1(phi = 0.3, sigma2 = 1), N = 1000, freq = 1)

#> Error in gen_gts(AR1(phi = 0.3, sigma2 = 1), N = 1000, freq = 1): unused argument (N = 1000)

avar(x, type = "to")

#> Error in as.vector(x): object 'x' not found

a = matrix(rep(0, 1000^2), nrow = 1000)
for (i in 1:1000){
  a[,i] = seq(from = 1 - i, length.out = 1000)
}
a.diag = diag(a)
a[upper.tri(a,diag=T)] = 0
a = a + t(a) + diag(a.diag)
covmat = 0.3^a
sapply(1:8, function(y){NOAV(2^y, covmat)})

#> [1] 0.523250000 0.350727081 0.203446388 0.108896684 0.056242028 0.028569436
#> [7] 0.014396823 0.007226438