Compute the latent processes parameters estimator based on the Allan Variance

Computes the Allan Variance

fit_av(x, qn = NULL, wn = NULL, rw = NULL, dr = NULL, type = NULL)

Arguments

x
A vec of time series observations or an avar object.

qn
A vec specifying on which scales the parameters of a Quantization Noise (QN) should be computed.

wn
A vec specifying on which scales the parameters of a White Noise (WN) should be computed.

rw
A vec specifying on which scales the parameters of a Random Wakk (RW) should be computed.

dr
A vec specifying on which scales the parameters of a Drift (DR) should be computed.

type
A string containing either "mo" for Maximal Overlap or "to" for Tau Overlap

Value

Fit_av A list that contains:

"process"The process selected

"theta_hat"The estimated parameter vector

Examples

# Load simts package
library(simts)

# Set seed for reproducibility
set.seed(999)

# Simulate time series
N = 100000
ts = gen_gts(N, WN(sigma2 = 2) + RW(gamma2 = 1))

# Maximal overlap
fit1 = fit_av(wn = 1:10, rw = 11:14)

#> Error in fit_av(wn = 1:10, rw = 11:14): argument "x" is missing, with no default